requirements.txt
fastapi
uvicorn[standard]
llama-stack-client==0.2.5
python-dotenv
